+++ ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: kkawano <kkawano@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/12 16:16:15 by kkawano           #+#    #+#             */
-/*   Updated: 2021/08/16 18:34:47 by kkawano          ###   ########.fr       */
+/*   Updated: 2021/08/17 22:04:58 by kkawano          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	if (!s)
 		return ;
